小叶子bot环境布置/requirement.txt
nonebot_plugin_clovers
clovers-AIchat
clovers-groupmate-waifu
clovers-leafgame
clovers-setu-collection
clovers-tabletop-helper